// frontend/src/backend.d.ts
export type Time = bigint;
export interface PaketMBGRecord {
    id: bigint;
    nama: string;
    jenis: Jenis;
    keterangan?: string;
}
export interface PendingBatchedDistribusi {
    idPaket: bigint;
    sasaranIds: Array<bigint>;
    keterangan?: string;
    statusDistribusi: DistribusiStatus;
    jumlahPaket: bigint;
    tanggalDistribusi: Time;
}
export interface BatchedDistribusiResult {
    gagal: Array<{
        error: string;
        sasaranId: bigint;
    }>;
    sukses: Array<DistribusiRecord>;
}
export interface DistribusiRecord {
    id: bigint;
    idPaket: bigint;
    idSasaran: bigint;
    keterangan?: string;
    statusDistribusi: DistribusiStatus;
    jumlahPaket: bigint;
    tanggalDistribusi: Time;
}
export type Jenis = {
    __kind__: "susuTambahan";
    susuTambahan: null;
} | {
    __kind__: "lainnya";
    lainnya: string;
} | {
    __kind__: "makananSiapSaji";
    makananSiapSaji: null;
} | {
    __kind__: "multivitamin";
    multivitamin: null;
} | {
    __kind__: "paketSembako";
    paketSembako: null;
};
export interface SasaranRecord {
    id: bigint;
    status: Status;
    alamat: string;
    nama: string;
    nomorIdentitas: string;
    kategori: Kategori;
    catatan?: string;
}
export interface UserProfile {
    name: string;
    role: string;
    email?: string;
}
export enum DistribusiStatus {
    pending = "pending",
    terdistribusi = "terdistribusi",
    dalamProses = "dalamProses",
    tidakTerkirim = "tidakTerkirim"
}
export enum Kategori {
    balita = "balita",
    tidakDitentukan = "tidakDitentukan",
    ibuMenyusui = "ibuMenyusui",
    ibuHamil = "ibuHamil"
}
export enum Status {
    aktif = "aktif",
    nonAktif = "nonAktif"
}